visualizations/requirements.txt
elastic-transport
elasticsearch
matplotlib-inline
numpy
pandas
plotly-express
SQLAlchemy
mysql-connector-python
streamlit


